refactor: Project architecture
Improved project architecture
- Rename utils files
- Create helpers folder

The `Helpers` folder should only contain functions that are dependent on the application.
The `Utils` folder should only contain functions that are not dependent on the application.
