+++ src/helpers/fetchCurrentUser.js
