+++ src/helpers/findIconByLabel.spec.js
@@ -1,16 +1,16 @@
 import People from 'cozy-ui/transpiled/react/Icons/People'
 
-import { getIconByLabel } from 'src/utils/getIconByLabel'
+import { findIconByLabel } from 'src/helpers/findIconByLabel'
 
-describe('getIconByLabel', () => {
+describe('findIconByLabel', () => {
   it('should return correct Icon if label is found', () => {
-    const result = getIconByLabel('national_id_card')
+    const result = findIconByLabel('national_id_card')
 
     expect(result).toBe(People)
   })
 
   it('should return nothing if label is not found', () => {
-    const result = getIconByLabel('fake_label')
+    const result = findIconByLabel('fake_label')
 
     expect(result).toBe(undefined)
   })
