+++ src/targets/browser/setupApp.jsx
@@ -2,8 +2,8 @@ import memoize from 'lodash/memoize'
 
 import { initTranslation } from 'cozy-ui/transpiled/react/I18n'
 
-import { getClient } from 'src/utils/client'
-import { getValues, initBar } from 'src/utils/bar'
+import { getClient } from 'src/helpers/client'
+import { getValues, initBar } from 'src/helpers/bar'
 import { RealtimePlugin } from 'cozy-realtime'
 
 /**
