+++ src/utils/hasNextvalue.js
@@ -1,1 +1,6 @@
+/**
+ * Check if there is a next value
+ * @param {number} curr - Current value
+ * @param {object} array - Array oc current value
+ */
 export const hasNextvalue = (curr = -1, array = []) => curr < array.length - 1
